requirements.txt
Flask==2.3.2
python-docx==0.8.11
fpdf==1.7.2
PyPDF2==3.0.1